mcp/gmail_server/requirements.txt
mcp[cli]
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
python-dotenv
beautifulsoup4
